refactor host entrypoint for testing
